requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-appconf==1.0.5
django-ckeditor-5==0.2.10
django-compressor==4.4
django-import-export==3.3.2
django-taggit==5.0.1
djangoql==0.17.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
ply==3.11
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyYAML==6.0.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
